refactor: reorganize cli utilities to declutter cli.py entrypoint
