+++ src/mkts_backend/cli_tools/add_watchlist.py
@@ -1,0 +1,71 @@
+from mkts_backend.config.logging_config import configure_logging
+
+logger = configure_logging(__name__)
+
+def add_watchlist(args: list[str]) -> None:
+
+    # Find the --type_id parameter
+    type_ids_str = None
+    for i, arg in enumerate(args):
+        if arg.startswith("--type_id="):
+            type_ids_str = arg.split("=", 1)[1]
+            break
+
+    if not type_ids_str:
+        print("Error: --type_id parameter is required for add_watchlist command")
+        print("Usage: mkts-backend add_watchlist --type_id=12345,67890,11111")
+        print("       mkts-backend add_watchlist --type_id=12345,67890,11111 --local")
+        return None
+
+    # Default to remote database, use --local flag for local database
+    remote = "--local" not in args
+
+    success = process_add_watchlist(type_ids_str, remote=remote)
+    if success:
+        print(f"Added {type_ids_str} to watchlist")
+    else:
+        print(f"Failed to add {type_ids_str} to watchlist")
+    exit()
+
+
+def process_add_watchlist(type_ids_str: str, remote: bool = False):
+    """
+    Process the add_watchlist command to add items to the watchlist.
+
+    Args:
+        type_ids_str: Comma-separated string of type IDs
+        remote: Whether to use remote database
+    """
+    try:
+        # Parse comma-separated type IDs
+        type_ids = [int(tid.strip()) for tid in type_ids_str.split(',') if tid.strip()]
+
+        if not type_ids:
+            logger.error("No valid type IDs provided")
+            print("Error: No valid type IDs provided")
+            return False
+
+        logger.info(f"Adding {len(type_ids)} items to watchlist: {type_ids}")
+        print(f"Adding {len(type_ids)} items to watchlist: {type_ids}")
+
+        # Call add_missing_items_to_watchlist with all type IDs at once
+        result = add_missing_items_to_watchlist(type_ids, remote=remote)
+
+        # Check if the operation was successful
+        if result.startswith("Error"):
+            logger.error(f"Failed to add items to watchlist: {result}")
+            print(f"Error: {result}")
+            return False
+        else:
+            logger.info(f"Successfully processed watchlist addition: {result}")
+            print(result)
+            return True
+
+    except ValueError as e:
+        logger.error(f"Invalid type ID format: {e}")
+        print(f"Error: Invalid type ID format. Please provide comma-separated integers. {e}")
+        return False
+    except Exception as e:
+        logger.error(f"Error adding items to watchlist: {e}")
+        print(f"Error: {e}")
+        return False
